1.1.0

+++ src/main/java/gs/teamup/bot/pojo/edge/CreateRoom.java
@@ -8,12 +8,10 @@ import lombok.Setter;
 /**
  * Created by thisno1 on 2016-01-20.
  */
-
 @Getter
 @Setter
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class CreateRoom {
-
     @JsonProperty("room")
-    int room;
+    Long room;
 }
